Initial Commit.
Right now I have a server.
It listens on port 1337, and accepts incoming connections.

The incomming connections type in lisp commands.
These are parsed by the lisp module into lists and atoms.

Then, there's a stop command that shuts down the server.

The code is rough, and the shutdown is not clean, but it works out ok.
Kind of.